+++ tests/unit/serve/executors/test_decorators.py
@@ -2,7 +2,6 @@ import functools
 
 import pytest
 
-from jina import Executor
 from jina.helper import iscoroutinefunction
 from jina.serve.executors import get_executor_taboo
 from jina.serve.executors.decorators import dynamic_batching, requests
@@ -66,65 +65,24 @@ def test_async_requests():
     assert iscoroutinefunction(getattr(fn_3, 'fn'))
 
 
-@pytest.mark.parametrize(
-    'inputs,expected_values',
-    [
-        (
-            dict(preferred_batch_size=4, timeout=5_000, max_batch_size=256),
-            dict(preferred_batch_size=4, timeout=5_000, max_batch_size=256),
-        ),
-        (
-            dict(preferred_batch_size=4, timeout=5_000),
-            dict(preferred_batch_size=4, timeout=5_000, max_batch_size=None),
-        ),
-        (
-            dict(preferred_batch_size=4, max_batch_size=256),
-            dict(preferred_batch_size=4, timeout=10_000, max_batch_size=256),
-        ),
-    ],
-)
-def test_dynamic_batching(inputs, expected_values):
-    class MyExec(Executor):
-        @dynamic_batching(**inputs)
-        def foo(self, docs, **kwargs):
-            pass
+def test_dynamic_batching():
+    @dynamic_batching()
+    def fn_2(*args, **kwargs):
+        pass
 
-    exec = MyExec()
-    assert exec.dynamic_batching['foo'] == expected_values
+    assert hasattr(fn_2, 'fn')
 
 
-@pytest.mark.parametrize(
-    'inputs,expected_values',
-    [
-        (
-            dict(preferred_batch_size=4, timeout=5_000, max_batch_size=256),
-            dict(preferred_batch_size=4, timeout=5_000, max_batch_size=256),
-        ),
-        (
-            dict(preferred_batch_size=4, timeout=5_000),
-            dict(preferred_batch_size=4, timeout=5_000, max_batch_size=None),
-        ),
-        (
-            dict(preferred_batch_size=4, max_batch_size=256),
-            dict(preferred_batch_size=4, timeout=10_000, max_batch_size=256),
-        ),
-    ],
-)
-def test_combined_decorators(inputs, expected_values):
-    class MyExec(Executor):
-        @dynamic_batching(**inputs)
-        @requests(on='/foo')
-        def foo(self, docs, **kwargs):
-            pass
+def test_async_dynamic_batching():
+    @dynamic_batching
+    def fn_2(*args, **kwargs):
+        pass
 
-    exec = MyExec()
-    assert exec.dynamic_batching['foo'] == expected_values
+    @dynamic_batching
+    async def fn_3(*args, **kwargs):
+        pass
 
-    class MyExec(Executor):
-        @requests(on='/foo')
-        @dynamic_batching(**inputs)
-        def foo(self, docs, **kwargs):
-            pass
-
-    exec = MyExec()
-    assert exec.dynamic_batching['foo'] == expected_values
+    assert hasattr(fn_2, 'fn')
+    assert not iscoroutinefunction(getattr(fn_2, 'fn'))
+    assert hasattr(fn_3, 'fn')
+    assert iscoroutinefunction(getattr(fn_3, 'fn'))
