+++ tests/unit/serve/executors/test_executor.py
@@ -9,15 +9,15 @@ import pytest
 import yaml
 from docarray import Document, DocumentArray
 
-from jina import Client, Executor, Flow, __cache_path__, requests
+from jina import Client, Executor, Flow, __cache_path__, dynamic_batching, requests
 from jina.clients.request import request_generator
 from jina.excepts import RuntimeFailToStart
+from jina.helper import random_port
 from jina.parsers import set_pod_parser
 from jina.serve.executors.metas import get_default_metas
 from jina.serve.networking import GrpcConnectionPool
 from jina.serve.runtimes.asyncio import AsyncNewLoopRuntime
 from jina.serve.runtimes.worker import WorkerRuntime
-from jina.helper import random_port
 
 
 class WorkspaceExec(Executor):
@@ -104,7 +104,7 @@ def test_executor_with_pymodule_path():
 
 def test_flow_uses_with_pymodule_path():
     with Flow.load_config(
-            '''
+        '''
     jtype: Flow
     executors:
         - uses: unit.serve.executors.dummy_executor.MyExecutor
@@ -115,13 +115,13 @@ def test_flow_uses_with_pymodule_path():
         pass
 
     with Flow().add(
-            uses='unit.serve.executors.dummy_executor.MyExecutor', uses_with={'bar': 123}
+        uses='unit.serve.executors.dummy_executor.MyExecutor', uses_with={'bar': 123}
     ):
         pass
 
     with pytest.raises(RuntimeFailToStart):
         with Flow.load_config(
-                '''
+            '''
             jtype: Flow
             executors:
                 - uses: jina.no_valide.executor
@@ -238,7 +238,7 @@ def test_executor_workspace(test_metas_workspace_replica_pods, shard_id):
 
 @pytest.mark.parametrize('shard_id', [None, -1], indirect=True)
 def test_executor_workspace_parent_replica_nopea(
-        test_metas_workspace_replica_pods, shard_id
+    test_metas_workspace_replica_pods, shard_id
 ):
     executor = Executor(
         metas={'name': test_metas_workspace_replica_pods['name']},
@@ -254,7 +254,7 @@ def test_executor_workspace_parent_replica_nopea(
 
 @pytest.mark.parametrize('shard_id', [0, 1, 2], indirect=True)
 def test_executor_workspace_parent_noreplica_pod(
-        test_metas_workspace_replica_pods, shard_id
+    test_metas_workspace_replica_pods, shard_id
 ):
     executor = Executor(
         metas={'name': test_metas_workspace_replica_pods['name']},
@@ -271,7 +271,7 @@ def test_executor_workspace_parent_noreplica_pod(
 
 @pytest.mark.parametrize('shard_id', [None, -1], indirect=True)
 def test_executor_workspace_parent_noreplica_nopea(
-        test_metas_workspace_replica_pods, shard_id
+    test_metas_workspace_replica_pods, shard_id
 ):
     executor = Executor(
         metas={'name': test_metas_workspace_replica_pods['name']},
@@ -406,7 +406,7 @@ def test_set_workspace(tmpdir):
         os.path.join(tmpdir, 'WorkspaceExec')
     )
     assert (
-            WorkspaceExec(workspace=str(tmpdir)).workspace == complete_workspace_no_replicas
+        WorkspaceExec(workspace=str(tmpdir)).workspace == complete_workspace_no_replicas
     )
 
 
@@ -451,10 +451,10 @@ def test_to_k8s_yaml(tmpdir, exec_type):
         with open(os.path.join(tmpdir, 'gateway', 'gateway.yml')) as f:
             gatewayyaml = list(yaml.safe_load_all(f))[-1]
             assert (
-                    gatewayyaml['spec']['template']['spec']['containers'][0]['ports'][0][
-                        'containerPort'
-                    ]
-                    == 2020
+                gatewayyaml['spec']['template']['spec']['containers'][0]['ports'][0][
+                    'containerPort'
+                ]
+                == 2020
             )
             gateway_args = gatewayyaml['spec']['template']['spec']['containers'][0][
                 'args'
@@ -560,7 +560,6 @@ def test_executors_inheritance_binding():
             pass
 
     class B(A):
-
         @requests(on='/index')
         def b(self, **kwargs):
             pass
@@ -577,3 +576,67 @@ def test_executors_inheritance_binding():
     assert set(C().requests.keys()) == {'/index', '/default', '_jina_dry_run_'}
     assert C().requests['/index'] == B.b
     assert C().requests['/default'] == A.default_a
+
+
+@pytest.mark.parametrize(
+    'inputs,expected_values',
+    [
+        (
+            dict(preferred_batch_size=4, timeout=5_000, max_batch_size=256),
+            dict(preferred_batch_size=4, timeout=5_000, max_batch_size=256),
+        ),
+        (
+            dict(preferred_batch_size=4, timeout=5_000),
+            dict(preferred_batch_size=4, timeout=5_000, max_batch_size=None),
+        ),
+        (
+            dict(preferred_batch_size=4, max_batch_size=256),
+            dict(preferred_batch_size=4, timeout=10_000, max_batch_size=256),
+        ),
+    ],
+)
+def test_dynamic_batching(inputs, expected_values):
+    class MyExec(Executor):
+        @dynamic_batching(**inputs)
+        def foo(self, docs, **kwargs):
+            pass
+
+    exec = MyExec()
+    assert exec.dynamic_batching['foo'] == expected_values
+
+
+@pytest.mark.parametrize(
+    'inputs,expected_values',
+    [
+        (
+            dict(preferred_batch_size=4, timeout=5_000, max_batch_size=256),
+            dict(preferred_batch_size=4, timeout=5_000, max_batch_size=256),
+        ),
+        (
+            dict(preferred_batch_size=4, timeout=5_000),
+            dict(preferred_batch_size=4, timeout=5_000, max_batch_size=None),
+        ),
+        (
+            dict(preferred_batch_size=4, max_batch_size=256),
+            dict(preferred_batch_size=4, timeout=10_000, max_batch_size=256),
+        ),
+    ],
+)
+def test_combined_decorators(inputs, expected_values):
+    class MyExecutor(Executor):
+        @dynamic_batching(**inputs)
+        @requests(on='/foo')
+        def foo(self, docs, **kwargs):
+            pass
+
+    exec = MyExecutor()
+    assert exec.dynamic_batching['foo'] == expected_values
+
+    class MyExecutor2(Executor):
+        @requests(on='/foo')
+        @dynamic_batching(**inputs)
+        def foo(self, docs, **kwargs):
+            pass
+
+    exec = MyExecutor2()
+    assert exec.dynamic_batching['foo'] == expected_values
